<TBBAS-1493> Fix method descriptions on permissions builder
