+++ core/coreobjects/include/coreobjects/permission_mask_builder.h
@@ -27,6 +27,17 @@ BEGIN_NAMESPACE_OPENDAQ
  */
 
 /*!
+ * @brief Enumeration of available access permissions
+ */
+enum class Permission : EnumType
+{
+    None = 0x0,    // The user has no permissions on the object.
+    Read = 0x1,    // The user can see and read an object.
+    Write = 0x2,   // The user can change or write to the object.
+    Execute = 0x4  // The user can execute an action attached to the object.
+};
+
+/*!
  * @brief A class which is responsible for creating a permission mask. This is a collection of Permission
  * values which are allowed or denied for a given group id. Permission mask is defined as a 64-bit integer,
  * where each bit corespond to a specific permission defined by Permission enum.
